增加传入年和周次获取该周周一的方法

+++ Taf.Core.Utility/Extensions/Extensions.Long.cs
@@ -30,7 +30,7 @@ namespace Taf.Core.Utility
         /// </param>
         /// <returns>
         /// </returns>
-        public static double MinutesToSeconds(this long obj) => obj * 60D;
+        public static long MinutesToSeconds(this long obj) => obj * 60;
         /// <summary>
         /// 间隔秒转小时
         /// </summary>
